Add QueryScope
